requirements.txt
config==0.4.2
docutils==0.14
Flask==1.1.1
Flask-Cors==3.0.7
glob2==0.6
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.21.0
google-cloud-core==1.0.3
google-cloud-datastore==1.10.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
html5lib==1.0.1
httplib2==0.14.0
jsonschema==2.6.0
matplotlib==3.0.2
numpy==1.15.4
numpydoc==0.8.0
oauth2client==4.1.3
oauthlib==3.1.0
pandas==0.23.4
path.py==11.5.0
pathlib2==2.3.3
prompt-toolkit==2.0.7
requests==2.21.0
requests-oauthlib==1.2.0
scikit-learn==0.20.1
scipy==1.1.0
tqdm==4.28.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
XlsxWriter==1.1.2
zict==0.1.3
